roll back a change, to make @sebersole's rebase easier
move IndexOrUniqueKeyNameSource back inside IndexBinder
